Basic unit tests for SNS position and velocity IK
This commit addes unit tests for:
- SNS position IK (all versions)
- SNS velocity IK (all versions)
- KDL position IK (pinv NR only)

It also addes a pair of functions  that directly create a KDL chain model
for Sawyer, which allows the unit tests to run without depending on external
repositories or URDF files.

sns_ik_data_utilities --> rng_utilities

This commit also turns on compiler warnings and treats them as errors.

Issues: #74