Adds a new functionality to enable a secondary goal (#96)
This commit adds new functionalities to vel/acc IK solvers
so that they can now take a configuration space task as a
secondary goal. A simple application of this feature will
be to have a nullspace bias which will help stabilizing
the nullspace motion as well as getting the solution pose
close to a desired pose. Unit tests are added as well.