Fixed a bug in SawyerModel defined for tests (#100)
This commit updates sawyer model defined for tests.
The previous model definition was incorrect, and
LIN_SOLVE_RESIDUAL_TOL had to be set very high (1e1)
because of that. Now, with the correct model definition,
the LIN_SOLVE_RESIDUAL_TOL is set to a reasonable number
(1e-8) which makes better sense.